% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_cases.R
\name{assert_cases}
\alias{assert_cases}
\title{Integrates logical statement asserts into the stpR object.}
\usage{
assert_cases(
  logic,
  stp_id = "99999",
  project_dictionary = get_project_dictionary(),
  project_directory = here::here()
)
}
\arguments{
\item{logic}{Statement that evaluates to a TRUE/FALSE.}

\item{stp_id}{Id of the stp_ob entity. Defaults to "99999".}

\item{project_dictionary}{Project dictionary. Defaults to the global dictionary.}

\item{project_directory}{Project directory. Defaults to the global directory based on \code{here::here()} call.}
}
\description{
Asserts that a statement is TRUE. If not, prints a warning and logs an alert.
}
\details{
If \code{throw_assert_errors} in the project_dictionary is \code{TRUE}, then an error is thrown instead of a warning.
If \code{throw_assert_errors} is \code{FALSE} and \code{console_output_global} is \code{TRUE} in the project_dictionary, then a warning is printed to the console.
}
\examples{
\donttest{

assert_cases(
  logic = qa_tbl \%>\%
    filter(trt == "Drug A" & stage == "T3") \%>\%
    nrow() == 0,
  stp_id = 'assert_trt_vs_stage'
)
}
}
