% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_distinct.R
\name{assert_distinct}
\alias{assert_distinct}
\title{Integrates asserts of uniqueness by key variables into the stpR object.}
\usage{
assert_distinct(
  data_df,
  group_by_vars = NULL,
  stp_id = "99999",
  project_dictionary = get_project_dictionary(),
  project_directory = here::here()
)
}
\arguments{
\item{data_df}{A data set to be evaluated for uniqueness.}

\item{group_by_vars}{A string or vector of strings representing fields within the data set to group by.}

\item{stp_id}{A character ID of the stp_ob entity. Defaults to "99999".}

\item{project_dictionary}{Project dictionary. Defaults to the global dictionary.}

\item{project_directory}{Project directory. Defaults to the global directory based on \code{here::here()} call.}
}
\description{
Asserts that rows in the data set are unique by one or more grouping variables. If not, prints a warning and logs an alert.
}
\details{
If \code{throw_assert_errors} in the project_dictionary is \code{TRUE}, then an error is thrown instead of a warning.
If \code{throw_assert_errors} is \code{FALSE} and \code{console_output_global} is \code{TRUE} in the project_dictionary, then a warning is printed to the console.
}
\examples{
\donttest{
temp_dup_tbl <- qa_tbl \%>\%
  dplyr::bind_rows(qa_tbl \%>\%
            dplyr::slice(1))

assert_distinct(
  data_df = temp_dup_tbl,
  group_by_vars = "col_id"
)

}
}
