% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_order.R
\name{assert_order}
\alias{assert_order}
\title{Integrate asserts of value order between two or more columns in a data set}
\usage{
assert_order()
}
\arguments{
\item{xyz}{abc}
}
\description{
If \code{throw_assert_errors} in the project_dictionary is \code{TRUE}, then an error is thrown instead of a warning.
If \code{throw_assert_errors} is \code{FALSE} and \code{console_output_global} is \code{TRUE} in the project_dictionary, then a warning is printed to the console.
}
\examples{
\donttest{


}
}
