% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stpr.R
\name{create_stp_ob}
\alias{create_stp_ob}
\title{Utility to create a new stp_ob object.}
\usage{
create_stp_ob(
  save_to_path = NULL,
  reset = FALSE,
  dir = here::here(),
  project_dictionary = get_project_dictionary()
)
}
\arguments{
\item{save_to_path}{If NULL, will attempt to save to the current_metadata_path in the project_dictionary.}

\item{reset}{If TRUE, will overwrite any existing stp_ob object at the save_to_path.}

\item{dir}{The directory to save the stp_ob object to, which by default is determined by the \code{here::here()} function.}

\item{project_dictionary}{The project_dictionary object, which by default is determined by the \code{get_project_dictionary()} function.}
}
\value{
stp_ob A list object that documents the current state of the project for evaluation and reporting.
}
\description{
Usually run at the beginning of processing a new data set.
}
\examples{
\donttest{

create_stp_ob(reset = TRUE)

}

}
