% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_map.R
\name{eval_map}
\alias{eval_map}
\title{Description}
\usage{
eval_map(
  data_df = NULL,
  from = NULL,
  to = NULL,
  std_proc_na = NULL,
  highlight = FALSE,
  issue = FALSE,
  notes = NULL,
  stp_id = NULL,
  save_artifact = TRUE,
  report = TRUE,
  perform_compare = TRUE,
  project_dictionary = get_project_dictionary(),
  project_directory = here::here(),
  summary = FALSE
)
}
\arguments{
\item{data_df}{a dataframe that contains the underlying data to be evaluated}

\item{stp_ob}{}
}
\description{
Description
}
