% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_summary.R
\name{eval_summary}
\alias{eval_summary}
\title{Description}
\usage{
eval_summary(
  summary_df = NULL,
  key,
  highlight = FALSE,
  issue = FALSE,
  notes = NULL,
  stp_id,
  save_artifact = TRUE,
  report = TRUE,
  perform_compare = TRUE,
  project_dictionary = get_project_dictionary(),
  project_directory = here::here()
)
}
\arguments{
\item{save_artifact}{create an artifact to be saved out (almost always true... only false if you just wanted to use this function to do the count() type check with benefit of console output control)}

\item{perform_compare}{case where you just want to highlight something, but didn't want to "enforce" any particular relationship}

\item{stp_ob}{}
}
\description{
Description
}
