% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{exec_view}
\alias{exec_view}
\title{Execute and view highlighted text.  Wrapper for View().}
\usage{
exec_view()
}
\description{
rstudio.prefs::use_rstudio_keyboard_shortcut("Ctrl+Shift+W" = "AmplioHelpers::exec_view")
}
