% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reporting_cases.R
\name{get_reporting_cases}
\alias{get_reporting_cases}
\title{Return cases relevant for reporting}
\usage{
get_reporting_cases(
  stp_ob = get_stp_object(),
  limit_to_highlight = FALSE,
  limit_to_issue = FALSE
)
}
\arguments{
\item{stp_ob}{stpR artifact object. Default is the current state of the stpR object.}

\item{limit_to_highlight}{boolean to limit returned values to those that are highlighted}

\item{limit_to_issue}{boolean to limit returned values to those that are flagged as issues}
}
\value{
a dataframe of the cases that are flagged for reporting
}
\description{
Based on the current state of the stpR object.  Currently, only returns those values that are flagged for reporting (i.e., report = TRUE).
In the future, this function will also allow returning "non-report" items that have updates associated with them as compared with the previous data set.
}
\examples{
\donttest{

  get_reporting_cases()

}
}
