% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stp_object.R
\name{get_stp_object}
\alias{get_stp_object}
\title{Returns stpR object.}
\usage{
get_stp_object(
  file_type = "current",
  dir = here::here(),
  project_dictionary = get_project_dictionary()
)
}
\arguments{
\item{file_type}{The type of stpR object to return, either "current" or "compare".  Default is "current".}

\item{dir}{The directory to save the stp_ob object to, which by default is determined by the \code{here::here()} function.}

\item{project_dictionary}{The project_dictionary object, which by default is determined by the \code{get_project_dictionary()} function.}
}
\value{
stp_ob A list object that documents the current state of the project for evaluation and reporting.
}
\description{
Utility to quickly pull stpR object from file.  Utilized by many of the strR functions for this purpose.
}
\details{
desc

details
}
\examples{
\donttest{

get_stp_object()

}
}
