% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_cases.R
\name{keep_cases}
\alias{keep_cases}
\title{Utility to filter cases from a data set that do not meet a specified condition.}
\usage{
keep_cases(
  data_df = NULL,
  stp_id = NULL,
  id_vars,
  non_id_vars,
  group_by_vars = NULL,
  keep_logic,
  highlight = FALSE,
  issue = FALSE,
  notes = NULL,
  save_artifact = TRUE,
  report = TRUE,
  perform_compare = TRUE,
  return_only_filtered = FALSE,
  project_dictionary = get_project_dictionary(),
  project_directory = here::here()
)
}
\arguments{
\item{data_df}{A data set to be evaluated and possibly filtered.}

\item{stp_id}{A unique identifier for this filter instance.}

\item{id_vars}{A string or vector of strings representing fields within the data set to use as identifiers.}

\item{non_id_vars}{A string or vector of strings representing fields within the data set to use as non-identifiers.}

\item{group_by_vars}{A string or vector of strings representing fields within the data set to group by.}

\item{keep_logic}{A string representing the logic to be used to filter the data set. Default is to keep cases where the logic evaluates to TRUE.}

\item{highlight}{A logical value indicating whether or not to highlight the filtered cases. Default is FALSE.}

\item{issue}{A logical value indicating whether or not to flag the filtered cases as an issue. Default is FALSE.}

\item{notes}{A string representing notes to be associated with the update artifact.}

\item{save_artifact}{A logical value indicating whether or not to save the update artifact. Default is TRUE.}

\item{report}{A logical value indicating whether or not to include the update artifact in the report. Default is TRUE.}

\item{perform_compare}{A logical value indicating whether or not to perform a comparison of the update artifact to the compare artifact. Default is TRUE.}

\item{return_only_filtered}{A logical value indicating whether or not to return only the filtered cases or the data set with these cases removed.  Default is FALSE}

\item{project_dictionary}{The project_dictionary object, which by default is output by the \code{get_project_dictionary()} function.}

\item{project_directory}{The project_directory object, which by default is output by the \code{here::here()} function.}
}
\value{
data frame. Depending on the return_only_filtered argument, either the data set with the filtered cases (default), or the filtered data itself.
}
\description{
This utility also creates an artifact of removed cases for reporting and comparison against separate data set processing.
}
\examples{
\donttest{

qa_tbl <- keep_cases(
  data_df = qa_tbl,
  stp_id = "age_filter",
  id_vars = "col_id",
  non_id_vars = "age",
  keep_logic = "age >= 18",
  notes = "Remove patients younger than 18 years old"
)

}

}
