% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{output_data_to_console}
\alias{output_data_to_console}
\title{Internal utility to output data frame and messages to the console based on several conditions.}
\usage{
output_data_to_console(
  df,
  header_text = NULL,
  cases_text = NULL,
  no_cases_text = "No cases to report",
  see_returned_object_text = "See returned object for further details"
)
}
\arguments{
\item{df}{The data frame to be output to the console.}

\item{header_text}{The text to be printed before the data frame.}

\item{cases_text}{The text to be printed before the data frame when there are cases to report.}

\item{no_cases_text}{The text to be printed when there are no cases to report.}

\item{see_returned_object_text}{The text to be printed when there are more than 9 rows in the data frame.}
}
\description{
Internal utility to output data frame and messages to the console based on several conditions.
}
