% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_cases.R
\name{replace_cases}
\alias{replace_cases}
\title{Utility to update individual column values within a data set.}
\usage{
replace_cases(xyz)
}
\arguments{
\item{xyz}{abc}
}
\value{
a dataframe with updated values
}
\description{
For tracking data update occurences.  Will be updated with code from the original.  Updated to include the ability to update multiple columns at once.
}
\examples{
\donttest{

  replace_cases()

}
}
