% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stp_filtered.R
\name{update_stp_filtered}
\alias{update_stp_filtered}
\title{Utility function to update the stp_ob "filtered" items artifacts with a new case.}
\usage{
update_stp_filtered(
  df_name,
  ref_ob,
  update_ob,
  stp_id_var,
  id_vars,
  group_by_vars,
  keep_logic,
  highlight,
  issue,
  notes,
  report,
  perform_compare,
  project_dictionary,
  project_directory
)
}
\arguments{
\item{df_name}{String name of the data frame that was being filtered.}

\item{ref_ob}{Reference object that was used to filter the data frame.}

\item{update_ob}{Update object that was used to filter the data frame.}

\item{stp_id_var}{String identifier for the case.}

\item{id_vars}{String vector of the id variables used to filter the data frame.}

\item{group_by_vars}{String vector of the group by variables used to filter the data frame.}

\item{keep_logic}{String logic used to filter the data frame.}

\item{highlight}{Boolean to indicate if the case should be highlighted. Defaults to FALSE.}

\item{issue}{Boolean to indicate if the case should be flagged as an issue. Defaults to FALSE.}

\item{notes}{String notes associated with the case.}

\item{report}{Boolean to indicate if the case should be flagged for reporting. Defaults to TRUE.}

\item{perform_compare}{Boolean to indicate if the case should be flagged for comparison. Defaults to TRUE.}

\item{project_dictionary}{The project_dictionary object.}

\item{project_directory}{The project_directory object.}
}
\description{
Used primarily internally by the stpR package.
}
