% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stp_mappings.R
\name{update_stp_mappings}
\alias{update_stp_mappings}
\title{Utility function to update the stp_ob "mapping" items artifacts with a new case.}
\usage{
update_stp_mappings(
  df_name,
  ref_ob,
  update_ob,
  from,
  to,
  std_proc_na,
  highlight,
  issue,
  notes,
  report,
  perform_compare,
  stp_id_var,
  project_dictionary,
  project_directory
)
}
\arguments{
\item{df_name}{String name of the data frame that was being filtered.}

\item{ref_ob}{Reference object that was used to filter the data frame.}

\item{update_ob}{Update object that was used to filter the data frame.}

\item{from}{string value of the column(s) in the data frame that represent the columns being processed in the current data set}

\item{to}{string value of the column(s) in the data frame that represent the columns being processed in the current data set.}

\item{std_proc_na}{A function that will be applied to the \code{from} column(s) to determine which values should be considered "standard processing" NA values.  This field is used in cases when processing values such as numbers, where it's not necessary to know all the "typically" processed values, but it is helpful to know the values that aren't readily processed by this standard processing as well as how they are handled in the current data set.}

\item{highlight}{Boolean to indicate if the case should be highlighted. Defaults to FALSE.}

\item{issue}{Boolean to indicate if the case should be flagged as an issue. Defaults to FALSE.}

\item{notes}{String notes associated with the case.}

\item{report}{Boolean to indicate if the case should be flagged for reporting. Defaults to TRUE.}

\item{perform_compare}{Boolean to indicate if the case should be flagged for comparison. Defaults to TRUE.}

\item{project_dictionary}{The project_dictionary object.}

\item{project_directory}{The project_directory object.}
}
\description{
Used primarily internally by the stpR package.
}
