% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stp_markdown.R
\name{create_stp_markdown}
\alias{create_stp_markdown}
\title{User is calling this function create rmarkdown}
\usage{
create_stp_markdown(
  include_reference_object = TRUE,
  include_update_object = FALSE,
  limit_to_highlight = FALSE,
  limit_to_issue = FALSE,
  rmarkdown_path,
  project_dictionary = get_project_dictionary(),
  project_directory = here::here()
)
}
\description{
User is calling this function create rmarkdown
}
