% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_map.R
\name{eval_fields}
\alias{eval_fields}
\title{Description}
\usage{
eval_fields(
  data_df = NULL,
  from = NULL,
  to = NULL,
  std_proc_na = NULL,
  highlight = FALSE,
  issue = FALSE,
  mapping_notes = NULL,
  update_notes = NULL,
  stp_id = NULL,
  artifact = TRUE,
  perform_compare = TRUE,
  project_dictionary = get_project_dictionary(),
  project_directory = here::here()
)
}
\arguments{
\item{stpr_ob}{}
}
\description{
Description
}
