% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fields.R
\name{eval_fields}
\alias{eval_fields}
\title{Description}
\usage{
eval_fields(
  data_df,
  from = NULL,
  to = NULL,
  excl_std_process = NULL,
  highlight = FALSE,
  issue = FALSE,
  summary_by = NULL,
  id = NULL,
  notes = NULL,
  artifact = TRUE,
  no_report = FALSE,
  compare = TRUE,
  stpr_ob = NULL
)
}
\arguments{
\item{stpr_ob}{}
}
\description{
Description
}
