% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fields.R
\name{eval_fields}
\alias{eval_fields}
\title{Description}
\usage{
eval_fields(
  data_df = NULL,
  from = NULL,
  to = NULL,
  std_proc_na = NULL,
  highlight = FALSE,
  issue = FALSE,
  notes = NULL,
  stp_id = NULL,
  artifact = TRUE,
  document = TRUE,
  compare = TRUE,
  stpr_project_dictionary = get_project_dictionary(stpr_project_dictionary),
  stpr_compare_object = get_compare_object(stpr_compare_object, stpr_project_dictionary)
)
}
\arguments{
\item{stpr_ob}{}
}
\description{
Description
}
