% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_summary.R
\name{eval_summary}
\alias{eval_summary}
\title{Description}
\usage{
eval_summary(
  summary_df,
  key,
  highlight = FALSE,
  issue = FALSE,
  notes = NULL,
  stp_id,
  save_artifact = TRUE,
  report = TRUE,
  perform_compare = TRUE,
  project_dictionary = get_project_dictionary(),
  project_directory = here::here()
)
}
\arguments{
\item{stpr_ob}{}
}
\description{
Description
}
