% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_dictionary.R
\name{get_project_dictionary}
\alias{get_project_dictionary}
\title{Read stpR project_dictionary from environment.}
\usage{
get_project_dictionary(
  filename = "project_dictionary.yaml",
  dir = here::here()
)
}
\arguments{
\item{filename}{The name of the project_dictionary file to read.  Default and standard is "project_dictionary.yaml".}

\item{dir}{The directory where the project_dictionary file is located.  Default is the current working directory.}
}
\value{
project_dictionary A list object that contains the project_dictionary.
}
\description{
By default, attempts to read the project_dictionary file from the current parent environment.
}
\examples{
\donttest{

  get_project_dictionary()

}
}
